# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_request.AttendanceRequestBulkDeleteRequest:
    properties:
      ids:
        items:
          type: string
        minItems: 1
        type: array
    required:
    - ids
    type: object
  github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_request.AttendanceRequestCreateRequest:
    properties:
      end_date:
        type: string
      evidence_file_id:
        type: string
      reason:
        type: string
      requested_overtime_hours:
        type: integer
      start_date:
        type: string
      type:
        allOf:
        - $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_model.AttendanceRequestType'
        enum:
        - sick
        - leave
        - extra_off
        - overtime
    required:
    - end_date
    - reason
    - start_date
    - type
    type: object
  github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_request.AttendanceRequestUpdateRequest:
    properties:
      end_date:
        type: string
      evidence_file_id:
        type: string
      reason:
        type: string
      requested_overtime_hours:
        type: integer
      start_date:
        type: string
      type:
        allOf:
        - $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_model.AttendanceRequestType'
        enum:
        - sick
        - leave
        - extra_off
        - overtime
    type: object
  github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_request.AttendanceRequestUpdateStatusRequest:
    properties:
      status:
        allOf:
        - $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_model.AttendanceRequestStatus'
        enum:
        - approved
        - rejected
    required:
    - status
    type: object
...
  github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response.AttendanceRequestResponse:
    properties:
      created_at:
        type: string
      employee_name:
        type: string
      end_date:
        type: string
      evidence_file_id:
        type: string
      evidence_file_url:
        type: string
      id:
        type: string
      reason:
        type: string
      requested_overtime_hours:
        type: integer
      review_note:
        type: string
      reviewed_at:
        type: string
      reviewed_by:
        type: string
      start_date:
        type: string
      status:
        $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_model.AttendanceRequestStatus'
      type:
        $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_model.AttendanceRequestType'
      updated_at:
        type: string
      user_id:
        type: string
    type: object
...
  github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.ActionSuccessResponse:
    properties:
      id:
        type: string
    type: object
  github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.DeleteSuccessResponse:
    properties:
      deleted_count:
        type: integer
    type: object
  github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.ErrorCode:
    enum:
    - BAD_REQUEST
    - VALIDATION_ERROR
    - AUTH_REQUIRED
    - AUTH_INVALID_CREDENTIALS
    - NOT_FOUND
    - FORBIDDEN
    - INTERNAL_SERVER_ERROR
    - DATABASE_ERROR
    - TOO_MANY_REQUESTS
    type: string
    x-enum-varnames:
    - BAD_REQUEST
    - VALIDATION_ERROR
    - AUTH_REQUIRED
    - AUTH_INVALID_CREDENTIALS
    - NOT_FOUND
    - FORBIDDEN
    - INTERNAL_SERVER_ERROR
    - DATABASE_ERROR
    - TOO_MANY_REQUESTS
  github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.ErrorSchema:
    properties:
      code:
        $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.ErrorCode'
      details: {}
      message:
        type: string
    type: object
  github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.Response-any:
    properties:
      data: {}
      error:
        $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.ErrorSchema'
      success:
        type: boolean
    type: object
  ? github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.Response-array_github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_AttendanceRequestResponse
  : properties:
      data:
        items:
          $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response.AttendanceRequestResponse'
        type: array
      error:
        $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.ErrorSchema'
      success:
        type: boolean
    type: object
...
  ? github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.Response-github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_AttendanceRequestResponse
  : properties:
      data:
        $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response.AttendanceRequestResponse'
      error:
        $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.ErrorSchema'
      success:
        type: boolean
    type: object
...
  ? github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.Response-github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common_ActionSuccessResponse
  : properties:
      data:
        $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.ActionSuccessResponse'
      error:
        $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.ErrorSchema'
      success:
        type: boolean
    type: object
  ? github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.Response-github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common_DeleteSuccessResponse
  : properties:
      data:
        $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.DeleteSuccessResponse'
      error:
        $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.ErrorSchema'
      success:
        type: boolean
    type: object
...
  github_com_afifudin23_absensi-king-royal-api_internal_model.AttendanceRequestStatus:
    enum:
    - pending
    - approved
    - rejected
    type: string
    x-enum-varnames:
    - AttendanceRequestStatusPending
    - AttendanceRequestStatusApproved
    - AttendanceRequestStatusRejected
  github_com_afifudin23_absensi-king-royal-api_internal_model.AttendanceRequestType:
    enum:
    - sick
    - leave
    - extra_off
    - overtime
    type: string
    x-enum-varnames:
    - AttendanceRequestTypeSick
    - AttendanceRequestTypeLeave
    - AttendanceRequestTypeExtraOff
    - AttendanceRequestTypeOvertime
...
host: localhost:8080
info:
  contact:
    email: afifudin23@example.com
    name: Afifudin
  description: REST API untuk manajemen absensi, karyawan, dan penggajian King Royal.
  license:
    name: MIT
  termsOfService: http://swagger.io/terms/
  title: Absensi King Royal API
  version: "1.0"
paths:
...
  /attendance-requests:
    delete:
      consumes:
      - application/json
      description: |-
        Menghapus satu atau beberapa pengajuan absensi sekaligus berdasarkan daftar ID.

        **Request Body:**
        - `ids` (array of UUID, required, min 1 item): Daftar ID pengajuan yang akan dihapus

        **Response 200:** Jumlah data yang berhasil dihapus
      parameters:
      - description: Daftar ID pengajuan yang akan dihapus
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_request.AttendanceRequestBulkDeleteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.Response-github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common_DeleteSuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.Response-any'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.Response-any'
      security:
      - BearerAuth: []
      summary: Hapus pengajuan absensi
      tags:
      - Attendance Requests
    get:
      description: |-
        Mendapatkan daftar semua pengajuan absensi dari seluruh karyawan. Akses hanya untuk admin.

        **Query Params (opsional):**
        - `status` (string): Filter status — `pending`, `approved`, `rejected`
        - `type` (string): Filter tipe — `sick`, `leave`, `extra_off`, `overtime`
        - `start_date` (string, YYYY-MM-DD): Filter dari tanggal ini
        - `end_date` (string, YYYY-MM-DD): Filter sampai tanggal ini

        **Response 200:** Array pengajuan beserta nama karyawan, status review, dan bukti
      parameters:
      - description: Filter status (pending/approved/rejected)
        in: query
        name: status
        type: string
      - description: Filter tipe (sick/leave/extra_off/overtime)
        in: query
        name: type
        type: string
      - description: Tanggal mulai (YYYY-MM-DD)
        in: query
        name: start_date
        type: string
      - description: Tanggal selesai (YYYY-MM-DD)
        in: query
        name: end_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.Response-array_github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_AttendanceRequestResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.Response-any'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.Response-any'
      security:
      - BearerAuth: []
      summary: Semua pengajuan absensi (Admin)
      tags:
      - Attendance Requests
    post:
      consumes:
      - application/json
      description: |-
        Membuat pengajuan kehadiran baru seperti cuti, sakit, lembur, atau libur tambahan.

        **Request Body:**
        - `start_date` (string, required, format YYYY-MM-DD): Tanggal mulai pengajuan
        - `end_date` (string, required, format YYYY-MM-DD): Tanggal selesai pengajuan
        - `reason` (string, required): Alasan pengajuan
        - `type` (string, required): Tipe — `sick`, `leave`, `extra_off`, `overtime`
        - `evidence_file_id` (UUID, opsional): ID file bukti (contoh: surat dokter untuk sakit)
        - `requested_overtime_hours` (integer, opsional): Jumlah jam lembur (khusus tipe `overtime`)

        **Response 201:** ID pengajuan yang baru dibuat
      parameters:
      - description: Data pengajuan
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_request.AttendanceRequestCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.Response-github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common_ActionSuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.Response-any'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.Response-any'
      security:
      - BearerAuth: []
      summary: Buat pengajuan absensi
      tags:
      - Attendance Requests
  /attendance-requests/{attendance_request_id}:
    get:
      description: |-
        Mendapatkan detail satu pengajuan absensi berdasarkan ID.

        **Path Param:**
        - `attendance_request_id` (UUID): ID pengajuan yang ingin dilihat

        **Response 200:** Data lengkap pengajuan (status, tipe, tanggal, alasan, bukti, reviewer, catatan review)
      parameters:
      - description: ID pengajuan
        in: path
        name: attendance_request_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.Response-github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_AttendanceRequestResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.Response-any'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.Response-any'
      security:
      - BearerAuth: []
      summary: Detail pengajuan absensi
      tags:
      - Attendance Requests
    put:
      consumes:
      - application/json
      description: |-
        Memperbarui pengajuan absensi yang masih berstatus pending. Pengajuan yang sudah diapprove/reject tidak bisa diubah.

        **Request Body (semua opsional):**
        - `start_date` (string, YYYY-MM-DD): Tanggal mulai baru
        - `end_date` (string, YYYY-MM-DD): Tanggal selesai baru
        - `reason` (string): Alasan baru
        - `type` (string): Tipe baru — `sick`, `leave`, `extra_off`, `overtime`
        - `evidence_file_id` (UUID): File bukti baru
        - `requested_overtime_hours` (integer): Jam lembur baru

        **Response 200:** ID pengajuan yang diupdate
      parameters:
      - description: ID pengajuan
        in: path
        name: attendance_request_id
        required: true
        type: string
      - description: Data yang diperbarui
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_request.AttendanceRequestUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.Response-github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common_ActionSuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.Response-any'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.Response-any'
      security:
      - BearerAuth: []
      summary: Update pengajuan absensi
      tags:
      - Attendance Requests
  /attendance-requests/{attendance_request_id}/status:
    patch:
      consumes:
      - application/json
      description: |-
        Admin menyetujui atau menolak pengajuan absensi karyawan. Akses hanya untuk admin.

        **Request Body:**
        - `status` (string, required): Status baru — `approved` atau `rejected`

        **Behavior:**
        - Jika diapprove, data absensi karyawan pada tanggal yang bersangkutan akan diperbarui otomatis
        - Reviewer dan waktu review dicatat otomatis

        **Response 200:** ID pengajuan yang diupdate statusnya
      parameters:
      - description: ID pengajuan
        in: path
        name: attendance_request_id
        required: true
        type: string
      - description: Status baru (approved/rejected)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_request.AttendanceRequestUpdateStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.Response-github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common_ActionSuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.Response-any'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.Response-any'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.Response-any'
      security:
      - BearerAuth: []
      summary: Setujui/tolak pengajuan (Admin)
      tags:
      - Attendance Requests
  /attendance-requests/me:
    get:
      description: |-
        Mendapatkan daftar semua pengajuan absensi milik pengguna yang sedang login.

        **Response 200:** Array pengajuan milik pengguna (status, tipe, tanggal, alasan, bukti, info review)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.Response-array_github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_AttendanceRequestResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.Response-any'
      security:
      - BearerAuth: []
      summary: Pengajuan absensi saya
      tags:
      - Attendance Requests
...
securityDefinitions:
  BearerAuth:
    description: 'Masukkan token JWT dengan format: **Bearer &lt;token&gt;** — contoh:
      `Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...`'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...